rename packages according to org.fairdatateam namespace
